requirements.txt
accelerate==0.33.0
badam==1.2.2
bitsandbytes==0.44.1
datasets==2.21.0
deepspeed==0.15.2
galore_torch==1.0
gradio==5.0.1
huggingface_hub==0.24.5
jieba==0.42.1
matplotlib==3.9.2
MoD==0.3.0
modelscope==1.18.1
nltk==3.9.1
numpy==2.1.2
optuna==3.6.1
packaging==24.1
pandas==1.5.3
peft==0.12.0
PyYAML==6.0.2
PyYAML==6.0.2
rdchiral==1.1.0
rdkit==2023.9.6
rouge_chinese==1.0.3
safetensors==0.4.5
torch==2.4.0
torch_geometric==2.6.1
transformers==4.44.0
trl==0.9.6
typing_extensions==4.12.2
vllm==0.6.2
